% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRdConsumer}
\alias{GetRdConsumer}
\title{GetRdConsumer}
\usage{
GetRdConsumer(keys, values)
}
\arguments{
\item{keys}{a character vector indicating option keys to parameterize the RdKafka::Consumer}

\item{values}{a character vector indicating option values to parameterize the RdKafka::Consumer. Must be of same length as keys.}
}
\description{
Creates an Rcpp::XPtr<RdKafka::KafkaConsumer>. For more details on options \link{https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md}
}
