% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRdProducer}
\alias{GetRdProducer}
\title{GetRdProducer}
\usage{
GetRdProducer(keys, values)
}
\arguments{
\item{keys}{a character vector indicating option keys to parameterize the RdKafka::Producer}

\item{values}{a character vector indicating option values to parameterize the RdKafka::Producer. Must be of same length as keys.}
}
\value{
a Rcpp::XPtr<RdKafka::Producer>
}
\description{
Creates an Rcpp::XPtr<RdKafka::Producer>. For more details on options \href{https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md}{Configuration}
}
