% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaBroker.R
\docType{data}
\name{Kafka Broker}
\alias{Kafka Broker}
\alias{KafkaBroker}
\title{KafkaBroker}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaBroker
}
\description{
TDB
}
\keyword{datasets}
