% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaBroker.R
\docType{data}
\name{KafkaBroker}
\alias{KafkaBroker}
\title{KafkaBroker}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaBroker
}
\description{
Brokers are servers that know the state
             of a Kafka cluster and participate in actions like
             creating / deleting topics and broadcasting messages
             to consumer groups.
}
\examples{
\donttest{
library(fRanz)

BROKER_HOST <- 'localhost'
BROKER_PORT <- 9092
TOPIC_NAME <- 'myTestTopic'
# KafkaBroker
broker <- KafkaBroker$new(host=BROKER_HOST, port=BROKER_PORT)

# KafkaProducer
producer <- KafkaProducer$new(brokers = list(broker))
producer$produce(topic = TOPIC_NAME,
                 key = "myKey",
                 value = "My First Message")
# Number of messages successfuly sent is returned
# [1] 1 


# KafkaConsumer
consumer <- KafkaConsumer$new(brokers = list(broker), groupId = "test", extraOptions=list(`auto.offset.reset`="earliest"))
consumer$subscribe(topics = c(TOPIC_NAME))
result <- consumer$consume(topic=TOPIC_NAME)

result
# Consumed messages are returned in a list(list(key,val)) format
# [[1]]
# [[1]]$key
# [1] "myKey"
#
# [[1]]$payload
# [1] "My First Message" 
}
}
\keyword{datasets}
