% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaConsumer.R
\docType{data}
\name{KafkaConsumer}
\alias{KafkaConsumer}
\title{Kakfa Consumer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaConsumer
}
\description{
TDB
}
\keyword{datasets}
