% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaProducer.R
\docType{data}
\name{KafkaProducer}
\alias{KafkaProducer}
\title{Kakfa Producer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KafkaProducer
}
\description{
A producer is an application that is responsible
             for publishing data to topics.
}
\references{
\href{https://kafka.apache.org/documentation/#intro_producers}{Apache Kafka docs - Producers}
}
\keyword{datasets}
