% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdSubscribe}
\alias{RdSubscribe}
\title{RdSubscribe}
\usage{
RdSubscribe(consumerPtr, Rtopics)
}
\arguments{
\item{consumerPtr}{a reference to a Rcpp::XPtr<RdKafka::KafkaConsumer>}

\item{Rtopics}{a character vector listing the topics to subscribe to}
}
\value{
the int representation of the librdkafka error code of the response to subscribe. 0 is good
}
\description{
A method to register a consumer with a set amount of topics as consumers. 
This is important so the broker can track offsets and register it in a consumer group
}
