% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KafkaConsume}
\alias{KafkaConsume}
\title{KafkaConsume}
\usage{
KafkaConsume(consumerPtr, numResults, timeout = 10000L)
}
\arguments{
\item{consumerPtr}{a reference to a Rcpp::XPtr<RdKafka::KafkaConsumer>}

\item{numResults}{how many results should be consumed before returning. Will return early if offset is at maximum}

\item{timeout}{the timeout in milliseconds. Default is 10000}
}
\value{
a list of length numResults with values list(key=key,value=value)
}
\description{
A method to that takes a consumer pointer and returns at most the specified number of results, unless the timeout is reached
}
